[hailgenetics] automate building and publishing of hailgenetics/python-dill
Previously python-dill lived both at gcr.io/hail-vdc/python-dill and gcr.io/hail-vdc/hailgenetics/python-dill
and hailgenetics/python-dill. This PR eliminates the first home because it is incongruous with
the other hailgenetics images. I also duplicated the Dockerfiles because that is the easiest
way to create multiple images in build.yaml.