[curl] pervasively use retries, report errors, follow redirects (#11183)
* [curl] pervasively use retries, report errors, follow redirects

I set a `.curlrc` file to ensure we use the right options by default. In a few places that are
intended to be run on a developer's laptop, I had to explicitly specify the necessary
options.

I replaced all the short flags with long flags, the important conversions to know:
- `-L` `--location`, follow redirects
- `-O` `--remote-name`, use the basename of the URL as the name of a local file in which to save the data
- `-s` `--silent`, do not print anything to stdout/stderr
- `-S` `--show-error`, when used with `-s`, shows errors on stderr
- `-f` `--fail`, return non-zero exit code if the response is 4xx or 5xx (normally the response would be printed / saved and curl will exit successfully)

* Update Dockerfile