[hailctl] Use a unique temp dir for starting browser in connect
gettempdir returns the system/Python temporary directory. If multiple
chromium processes are launched with the same user data directory, the
the process is reused. Because of this, new proxy settings won't be
picked up if the directory is reused. So we use a random identifier for
the chromium data directory.

Co-Authored-By: Dan King <daniel.zidan.king@gmail.com>