[batch] ensure cost is up-to-date when a batch is marked complete (#11406)
* [batch] ensure cost is up-to-date when a batch is marked complete

Closes hail-is/hail-production-issues#3

Details [here](https://github.com/hail-is/hail-production-issues/issues/3).

The high-level takeaway is that if the system is moving quite fast MJS
can happen *after* MJC. When MJC happens first, there are not yet any
resources. As a result, the state of a batch is observable as
`complete` *before* the resources are present. A lack of
resources implies an incorrect total cost for a batch. Linking
resource insertion with MJC ensures a batch may be complete if-and-only-if
the cost of the batch is accurate.

* export the Transaction type too

* fix for black

* missing awaits