[java-services] add new transient error (#11402)
See https://ci.hail.is/batches/532603/jobs/112.

We encounter a `is.hail.relocated.com.google.cloud.storage.StorageException` which is caused by a
`com.google.api.client.http.HttpResponseException`. The latter exception is not currently
considered a transient error. This PR changes isTransientError to recognize `HttpResponseException`
as a transient error.