[query] Lower TextTableWriter (#11323)
* [query] Lower TextTableWriter

To do this add an intermidiate abstract PartitionWriter,
SimplePartitionWriter, and extend it with TextTablePartitionWriter.

SimplePartitionWriter manages the output stream for its subclasses. The
subclasses need to implement consumeElement, and optionally preConsume
and postConsume.

TextTablePartitionWriter handles the delimiter delimited writing per
element item and the line delimited writing per element of the stream.
It will also optionally write the header if ExportType.PARALLEL_HEADER_IN_SHARD
has been requested.

TextTableFinalizer is the 'MetadataWriter' (name change pending), for
TextTableWriter. That's where the header is written for
PARALLEL_SEPARATE_HEADER and CONCATENATED, and files are merged for
CONCATENATED.

There are currently a few idosyncracies that don't replicate bug-for-bug
compatibility with the non-lowered version.

    1. We avoid the use of fs.copyMerge, as such, we don't check for the
       existence of the _SUCCESS file (even though we create it).
    2. ExportType.PARALLEL_COMPOSABLE is unsupported.
    3. We rely on the temporary file cleaner to clean up the files
       created for CONCATENATED export, rather than deleting them
       explicitly.

* fix TableTextWriter.canLowerEfficiently

Set it false for ExportType.PARALLEL_COMPOSABLE

* clean up temporary files in CONCATENATED export

* use fs in table export test

* use fs copy in test_import_table_force_bgz

* fix

* fix (can't use open with the ServiceBackend)

* final fix (:pray:)