[batch] wait unless we have more work (#11384)
* [batch] wait unless we have more work

The return value of these functions indicates if the containing loop
should wait or if we should immediately re-call the function. This
is intended to be used to allow functions which *know* they have more
work to eagerly invoke themselves again.

The use of this variable seems to have been changed to basically always
eagerly re-run during the Azure work.

This change restores the original behavior:
1. Do not wait in job private if we saw 300 records (seems likely there were
   301 or more records in the db).
2. Do not wait in pool scheduler if we exhaust a user's share. I do not
   fully follow the pool scheduler's logic. There might be something
   smarter we can do. I think we should really only re-call if we believe
   the db contains more ready jobs and we have available cores.

* Update pool.py