[batch] Add nonpreemptible pools (#11594)
* [batch] Add nonpreemptible pools

I don't think I need extra tests for this because the service tests will do so. However,
they won't use the highmem or highcpu pools. We can add tests for these, but it will add
time to the batch tests. We might want to split the batch tests up more in that case.

* fix

* fix

* fix