[query/service] simplify ServiceBackend by abstracting over ExecuteContext
Previously, I read the tmpdir, billingProject, and remoteTmpdir arguments
for each comand in the branch of the switch for that command. Then, I called
a method for that command, passing the three aformentioned arguments. The
method then constructs an execute context. Every method constructed the
context in the exact same way. This was just naughty copy paste job on my
part.

I rectify that here by abstracting over the creation of the execute
context and moving the reading of the three execute context relevant
parameters to outside of the switch.