[lowering] lower vcf export  (#11450)
* Remove auto gzip behavior in python FS

* [lowering] lower vcf export

Re-implement export vcf in generated code.

There is a fair amount of 'duplicated' code here between table export
and vcf export, however, I belive this to be fine. We can always
refactor VCFPartitionWriter to be a subclass of SimplePartitionWriter,
but that would require a little special casing as VCF export needs
access to the column values and SimplePartitionWriter assumes such
a thing is not necessary.

As far as VCF export itself, we simply duplicate the logic present in
ExportVCF but with generated code.

* last xfail