[query/service] use error id to raise user-friendly errors (#11624)
* [query/service] use error id to raise user-friendly errors

The main change is to the communication protocol between the client and
the driver and between the driver and the worker.

In main, both the driver and the client send messages back to the client
and driver (respectively) by writing to a file in cloud storage. In both
cases, the file (in main) has one of these two structures:

    0x00                  # is_success (False)
    UTF-8 encoded string  # the stack trace

    0x01                  # is_success (True)
    UTF-8 encoded string  # JSON message to send back to the client or driver

In this PR, the success case does not change. The failure case becomes:

    0x00                  # is_sucess (False)
    UTF-8 encoded string  # short message
    UTF-8 encoded string  # expanded message
    4-byte signed integer # error id

The service backend in Python changes to read this and raise the right error
if an error id is present.

I also uncovered three unrelated problems that are fixed in this PR:
1. PlinkVariant needs to be serializable because it is broadcasted.
2. We open an input stream in LoadPlink which ought to be closed, but there is no mechanism to do so in the ServiceBackend. I just ignore it for now. cc: @tpoterba, I'm not sure what the right answer is here.
3. Two uses of the broadcasted file system that should use the ExecuteContext's file system.

* fix issues

* new test passes

* another passing test!