[query/service] even more retries (#11709)
This retries all the basic RPC methods of GoogleStorageFS. I saw a transient
error arise from an `open` in generated code. These changes would have
retried the `open` and not failed the job (1 of 50,000 jobs :( ). However,
I am still not sure how to retry an error during actual writing (as opposed
to opening the OutputStream). I think we have to retry the whole partition.
See further discussion in Zulip.