[query] Fix high memory usage VDS converters after splitting multiallelics
The `to_dense_mt` and `to_merged_sparse_representation` methods required
localizing into an array all of the split rows at each locus when the
variant data is on the right side of the join. We flip the order so that
the variant data are on the left and the reference data, known to be
locus-distinct, are on the right.

I also fix a few stream misparameterizations that led to a slightly
higher (but similarly scaling) memory usage.