[httpx] use timeouts and remove footgun
`httpx.ClientSession` did not include the logic to set up the timeout. I do
not recall why I had a factory function distinct from the class itself. That
was a mistake. I leave the factory function because it is used prevasively.
However, now, allocating the class itself is also OK.