[query/shuffler] bind CDA to ensure it is run once
This seems to me like a footgun. Perhaps the methods in LowerTableIR should all
bind the CDA before handing it to the body? That will substantially change the
structure of some of the writers which assume that they can wrap the execution
of CDA in a `RelationalWriter.scoped`. I took that approach at first and worried
that it was too large a change.

This change reduces one particular pipeline I was investigating from ~20 stages
to ~5 stages.