[query] import_table: simplify code and improve speed (#11782)
* [query] import_table: simplify code and improve speed

CHANGELOG: `hl.import_table` is up to twice as fast for small tables.

I mostly abstracted things and simplified code. One big change is optimizing
for the single file, no filters case in which we need not scan for the first
extant row, that row *must* be in the first partition, if it exists at all.

* handle empty files