[query] Fix latent bug in ArrayElementAggregator copy (#11805)
I spruced up the entire method, but the bug was in the original
L153. One of the Code=>Value changes in the last couple of months
added the memoize to eltOffset. This was a ticking time bomb --
the usage of eltOffset inside `seq` must be evaluated inline, since
the value of `idx` changes between iterations of the loop where the
code argument to `seq` is located.

I suspect this has worked until now because `idx` is 0 every time
this method is called in typical runtimes. It exploded for the
buffered aggregator PR, where we are calling the copy with a stale
value of `idx` presumably equal to the length of the array, so we're
reading memory off the end of the array allocation.