[query] Allow tracking of the full deserialized byteSize of a partition
Add an extra (currently unused) flag on PartitionNativeWriter and an
extra field on its return type. When the flag is true, we compute the
sum of the byte sizes of every row and return it as the extra field
`partitionByteSize`.

We will use this to compute branch factors in LowerDistributedSort and
depending on computational cost, possible turn this on by default to
allow more optimization around deserializing/processing more than one
row at a time.