remove libhail (#11831)
Some repository cleanup. The libhail experiment has lain dormant for
a long time now. It is probably not the direction that we will take the
hail project in, it was a useful experiment but at this point, it is
extra bulk in the repository that will remain in the project history,
but need not be present in the current worktree.