[copy] finally eliminate all stalls from the copy tool (#11778)
* [copy] finally eliminate all stalls from the copy tool

No changelog because the copy tool is not properly public, though
maybe it can be public now.

The main problem was that we were saturating our network bandwidth
with data we did not need. Why? As far as I can tell, `release` on
an `aiohttp.ClientResponse` is insufficient to stop the receipt of
more bytes.

I realized one proper fix was to simply tell the cloud storage vendors
how much data we wanted to receive. That is the change to `open_from`
which I have made pervasively to all clouds.

I also changed `release` to `close` because that seems more correct.
I do not understand why we only `release`d it before.

* address comments

* fix for azure and test

* fixes

* fix

* fixes for local too

* fix length=0;

* fix

* fix gcs open_from name

* fixes

* add missing abstract methgod

* fix pylint

* fix

* fixes

* n to offset

* fixes

* notimplemented abstract methods

* fix

* fix

* fix