[query] Remove unexpected files from parallel text exports
This fixes an issue introduced by #11910, where extra part files may be
created by speculative execution. This is an issue for parallel export
where we did not delete any files.

Now we list the output directory, and compare it to the proper list of
files. The paths that are not in the list of 'good' paths are deleted.