[query] Specialize GoogleStorageFS.copy (#11922)
Issue the [rewrite] request. To save on reading and writing entire
blobs, as well as transferring data to the machine running the copy
rather than letting google storage move the data.

This operation should be effectively free when copying from one location
in a bucket to a different location in the same bucket, as the linked
docs say, when source and destination have the same location and storage
class, the blob is copied in a single request and returns immediately.

[rewrite]: https://cloud.google.com/storage/docs/json_api/v1/objects/rewrite