[fs] not FileNotFoundError for glob patterns (#11905)
* [fs] not FileNotFoundError for glob patterns

Currently, if we have a file structure like:

    a/
        b/
    aa/
        bb/

A glob pattern like `*/b` will raise FileNotFoundError beacuse
we try to list the file or folder named "b" inside `aa`. We should
not error. We should return `[a/b]`.

It is insufficient to avoid FileNotFoundError altogether because
the Hadoop API treats paths without globs differently. In particular,
listing the path `aa/b` should raise an error.

This change fixes behavior in the first case and treats the second
case explicitly.

* fix