[QoB] protect reference download with a lock
The tests use multiple threads which can race to download the references. This is a bit
of a blunt fix. In particular, this is not an asyncio-friendly lock (because I need
thread safety, which asyncio.Lock does not provide). In general, users should not try
to initialize hail multiple times in different coroutines in the same thread.