Create project skeleton for native hail query compiler/runtime
We have chosen to prototype a new hail query compiler and runtime using
the LLVM project's MLIR framework for query compilation infrastructure.
This creates a simple project skeleton that will serve as a jumping off
point for the work. The README should be enough to get it built. There
is an implementation of the opt tool for investigating IRs, and
a completely empty dialect.

I chose to rewrite this rather than copy over [hail-is/mlir-hail] as
breaking API changes from LLVM/MLIR 14 to 15 made it very difficult to
update that repo. The difficulties in updating LLVM versions may prove
to be a source of pain going forward, as this will be a pretty deep
dependency of the system.

[hail-is/mlir-hail]: https://github.com/hail-is/mlir-hail