[query] Fix lowering of aggs to preserve free variables in init args (#12092)
* [query] Fix lowering of aggs to preserve free variables in init args

CHANGELOG: Fixed compiler bug: "NoSuchElementException: Ref with name __iruid_X could not be resolved in env"

This is a bad change to a worse problem. The right solution is to redesign the IR
so that implicit init eval scopes don't exist -- but the right solution is hard to justify
going off to do right now.

This change patches the Extract.scala lowering logic to track the variables bound inside a
lowered IR and find the highest node that provides all necessary free variables. This change
still makes assumptions about the structure of the IR -- namely, it is still invalid to write
an IR like:

```
MakeTuple
  Let
    initBinding1
    <something>
    ApplyAggOp with ref to initBinding1
  Let
    initBinding2
    <something2>
    ApplyAggOp with ref to initbinding2
```

However, this fix resolves the case where init args reference a high single binding chain, as
in the test added.

* no sharing

* move requiredness

* fix

* fails