[ggplot] avoid circularity in types
I love simpler solutions that leverage mutation, but in this case,
`ggplot.py` and `geom.py` are circularly dependent if we want to
use `GGPlot` in `geom.py`. I return a Boolean indicating if this
geom makes the plot static. I do not love it because there is little
indication of what that return value means, but I do not see another
simple solution.