[ggplot] avoid weird dataframe error (#12337)
I do not fully understand why, but I was getting errors about df.columns
not accepting string indexes:

```
IndexError: only integers, slices (`:`), ellipsis (`...`), numpy.newaxis (`None`) and integer or boolean arrays are valid indices

```

I think `df.columns` is not intended to let us get the Series for a
column from the DataFrame. I think we are supposed to use `df` directly
for this purpose, which is what I do here.

I also noticed an unfortunate issue where `` is false-y so it is
overlooked by the `or` in favor of the default value. It would be nice
to have some kind of lazy `None` coalescing operator in Python :/.