[vds combiner] Save dataset type in combiner state
The dataset type should not change with each stage in the combiner, so
we are able to assert the type of the combiner datasets at each stage
and prevent round trips to the backend during reading. As such, we can
remove a call to read_multiple_matrix_tables which requires spark.

This sets up more changes so that we can switch on the backend and run
the combiner in query on batch.