[batch/qob] propagates errors from failed QoB worker jobs without result files
Currently, when a QoB worker job fails in a way that prevents it from
writing a result file with the error's stack trace in it, the Batch
worker that started the job tries to get the result file anyway, and
when it fails, raises a 404. This change retrieves the stack trace from
the QoB job and raises an error with the stack trace included.