[query][qob] simplify QoB error handling and fix flaky test
The root issue here was that sometimes exc.args[0] was a string and sometimes it was a dict. When
it was a string the `in` condition worked fine. When it was a dict, it was looking at the keys of
the dict and not finding the error message (which is buried under a few layers).

The code was unnecessarily complex. I reworked the yaml printer to be simpler and work for any
multiline string. I removed the regular expression that was used to discover the worker batch
when the worker jobs were in a different batch from the driver jobs. I remove all specialized
debugging information in favor of the general `debug_info` methods on `Batch` and
`ServiceBackend`. I also have two clear error cases: if the driver does not write its output
file, then something went horribly wrong. We dump all the debug info. If we do not receive
valid JSON from the driver, again, something went horribly wrong. We dump all the debug info.
The only remaining exceptional case is an error purposely serialized by the QoB driver to us
(with or without an error id).