[query/ggplot] Updates faceting to respect legend grouping
CHANGELOG: Added support for `scales`, `nrow`, and `ncol` arguments, as
well as grouped legends, to `hail.ggplot.facet_wrap`.

This change brings the `facet_wrap` interface more in line with
[the `ggplot2` implementation](https://ggplot2-book.org/facet.html#facet-wrap)
by defaulting to show axes (with tick marks) for all facets, as well as
providing `scales`, `nrow` and `ncol` arguments that the user can use to specify
whether the scales should be the same across facets and how many rows or columns
to use.

It also updates the faceting code to avoid duplicating legend entries when
[grouped legends](https://github.com/hail-is/hail/pull/12254) are used, and to
disable interactivity accordingly.