[query] Fix memory leak of relational keys in native writers (#12574)
In order to track partition bounds and distinctness, we report the first
and last seen keys when writing (matrix) tables. Previously we were
copying the last seen key into the partition region. This is incorrect
as the partition region has a lifetime of the entire partition and
cannot be cleared, leaking memory. Fix this by giving the last seen key
its own region that can be cleared before a new last seen key is saved.