[query][qob][hailtop/fs] Expose modification and creation time. (#12571)
* [query][qob][hailtop/fs] Expose modification and creation time.

Fixes #12540.

CHANGELOG: When using Query-on-Batch, hl.hadoop* methods now properly support creation and modification time.

Creation time is supported by modern Linuxes but only through a new statx API which is not
exposed by the Python standard library. There is a 0.1 version library from 2021 which exposes
statx including the "birth time". I chose to raise an exception for now.

Each cloud does support a "modification time" but it generally refers to changes to metadata or is just the creation time:
- https://cloud.google.com/storage/docs/json_api/v1/objects#resource (see updated)
- https://docs.aws.amazon.com/AmazonS3/latest/API/API_GetObject.html#API_GetObject_ResponseSyntax ("Last-Modified", is always creation time)
- https://docs.aws.amazon.com/AmazonS3/latest/API/API_HeadObject.html#API_HeadObject_ResponseSyntax same as above
- https://learn.microsoft.com/en-us/rest/api/storageservices/get-blob (see Last-Modified and x-ms-creation-time)

* test fixes

* unused importg

* lints

* fix

* pacify mypy

* fix type incompatibility in test