[query] improve the memory client
Currently, the memory client buffers the entire output in memory which is likely
to cause OOMs. For reasons that are not entirely clear to me, sometimes these OOMs
get muffled by our system and instead lead to non-termination. I vaguely remember
this happening before with `using`. I suspect there is something somewhat subtle
wrong with that method, but I am not certain.

Anyway, there are four big changes here:
1. Do not buffer the entire request body in memory when writing to memory.
2. Because of (1) we have to pull retry behavior all the way up to the top-level where we know how to recreate the body.
3. Because of (2) it is easier to provide a `write(url)(writerFunction)` style API, which I do here.
4. Again, because of (2), and because I want to preserve the file-object-like interface, I added a somewhat funky anonymous class which uses a second thread to facilitate the movement of data written into the OutputStream returned by `create` into the OutputStream of the HTTP connection.

Point (4) probably bears more explanation. The root issue is the bad Apache HTTP Client interface.
Instead of `request` returning an OutputStream, it takes an "entity". An entity knows how to write
itself into the OutputStream of an HTTP request. This works fine if the "writer" code is pased as
a function (as in my new `write` method), but that does not work if the control flow looks like:

    f = create(...)
    f.write(...)
    f.close()

We avoid this limited API by initiating the request in a second thread which will eventually block
waiting to receive data from a PipedInputStream. That PipedInputStream produces the data written to
a PipedOutputStream. The `create` call returns a positioned OutputStream which just writes data into
the PipedOutputStream and handles cleaning up the thread when it is closed.

In a multi-core system, network requests should proceed in parallel to the client code. In a
single-core system, the written data will buffer until `close` is called which will definitely yield
control to the other thread.