[query] add `hl.utils.genomic_range_table` (#12679)
* [query] add `hl.utils.genomic_range_table`

CHANGELOG: In Query on Batch, `hl.balding_nichols_model` is slightly faster. Also added `hl.utils.genomic_range_table` to quickly create a table keyed by locus.

It has grated on me for a while that `hl.balding_nichols_models` requires a whole pass to verify it
is sorted even though it is plainly so. This change introduces the necessary infrastructure to
convince Hail of that fact.

* pylint

* missing import

* fix tests

* add to __init__.py

* better genomic range test tables