[ci] Fix default namespace record for dev namespaces (#12756)
Stacked on #12755, this shouldn't have an impact on production, but is
just meant to replace `default` in test CI databases with the actual
name of the namespace that the CI is in. Should mimic [the prior
migration](https://github.com/hail-is/hail/blob/main/ci/sql/active-namespaces.sql)
just parametrized by namespace.